+++ ffs/types.go
@@ -148,7 +148,7 @@ func (s StorageConfig) WithColdFastRetrieval(enabled bool) StorageConfig {
 
 // WithColdStartDealOffset sets the maximum time in epochs a new deal must
 // be active on-chain.
-func (s StorageConfig) WithColdStartDealOffset(offset int) StorageConfig {
+func (s StorageConfig) WithColdStartDealOffset(offset int64) StorageConfig {
 	s.Cold.Filecoin.DealStartOffset = offset
 	return s
 }
@@ -356,7 +356,7 @@ type FilConfig struct {
 	// deadline to new deals being active on-chain. This value might influence
 	// if miners accept deals, since they should seal fast enough to satisfy
 	// this constraint.
-	DealStartOffset int
+	DealStartOffset int64
 }
 
 // Validate returns a non-nil error if the configuration is invalid.
